requirements.txt
streamlit
pandas
requests
google-genai
beautifulsoup4
xmltodict
